Re-implement and adapt type sharding to Grakn Core 1.5.x (#5412)
## What is the goal of this PR?

Type sharding is a feature which is intended to maintain good query performance when working with large graph.  
 The main aim of type sharding is to alleviate a well-known graph problem known as the "supernode" problem. A supernode refers to a vertex which has many incidental edges. Queries that need to touch such a vertex will be slow. The bigger the graph is, the more likely it is for a supernode to exist in the graph.

In Grakn, a supernode can exist when there exists a single type with millions of instances. In this scenario, the type vertex is the supernode as it will have millions of edges to each and every instance.

With type sharding, the type vertex will be connected to several "shards", each of which connected to a portion of the instances.

Type sharding is performed when a transaction is committed.

## What are the changes implemented in this PR?
- Perform type sharding on `TransactionOLTP::commit` when the threshold is reached. Determining when the next time we should shard is done using checkpoints. The checkpoints are persisted in the graph (see `TypeShardCheckpoint` for the implementation).
- Added tests in `TransactionOLTPIT` which test shard creation and threshold configuration
- Refactor: moved `TransactionOLTPIT` from the `kb` package into the more appropriate `session` package.