Add a check to avoid traversing the tree of all concepts to write's dependants if no concepts have been inferred in the first place (#5418)
## What is the goal of this PR?
We incur a substantial performance hit by searching the dependency tree of every concept that is written/queried in a `match...insert`. This can entirely be avoided if there are 0 inferred concepts live in this Transaction. The PR implements this slightly conservative check that should be a large help with data loading jobs using `match...insert` without any inference.

## What are the changes implemented in this PR?
* Check for the existence of any inferred concepts before actually searching the dependency tree of every concept that matched or inserted.